control-plane/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
paramiko==3.4.0
kubernetes==29.0.0
